fn/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
bs4
requests
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
transformers